#PV2-3472
Add mapping for LearningStartDate in RequiredPaymentsProfile

This commit introduces a mapping for the `LearningStartDate` property in the `RequiredPaymentsProfile` class. The property is now included in the mapping configuration for `RequiredPayment` to `CalculatedRequiredIncentiveAmount` and `CalculatedRequiredLevyAmount`. The previous mapping for `LearningStartDate` has been removed, reflecting a change in how this property is managed in the mapping process.